feat(backend): Create class library Presentation for controllers
