+++ MotorcycleCompany/LoggerService/LoggerManager.cs
@@ -12,8 +12,7 @@ namespace LoggerService
         {
                 
         }
-        public void LogDebug(string message) => logger.Debug(message);
-        
+        public void LogDebug(string message) => logger.Debug(message);       
 
         public void LogError(string message) => logger.Error(message);
