+++ MotorcycleCompany/MotorcycleCompany.sln
@@ -17,7 +17,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Service.Contracts", "Servic
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Service", "Service\Service.csproj", "{87CDC9BC-CF00-46C4-BF4B-E0F6BA8F3BE4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shared", "Shared\Shared.csproj", "{E220ACD5-E1F7-4BA1-867A-B80F30647AF7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Shared", "Shared\Shared.csproj", "{E220ACD5-E1F7-4BA1-867A-B80F30647AF7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Presentation", "Presentation\Presentation.csproj", "{12489DCB-CAD0-4A14-B72C-4EBD9CFF7966}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -57,6 +59,10 @@ Global
 		{E220ACD5-E1F7-4BA1-867A-B80F30647AF7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E220ACD5-E1F7-4BA1-867A-B80F30647AF7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E220ACD5-E1F7-4BA1-867A-B80F30647AF7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{12489DCB-CAD0-4A14-B72C-4EBD9CFF7966}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{12489DCB-CAD0-4A14-B72C-4EBD9CFF7966}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{12489DCB-CAD0-4A14-B72C-4EBD9CFF7966}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{12489DCB-CAD0-4A14-B72C-4EBD9CFF7966}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
