--- MotorcycleCompany/MotorcycleCompany/Controllers/CityController.cs
@@ -1,62 +1,0 @@
-﻿using Entities.Models;
-using Microsoft.AspNetCore.Mvc;
-using Service.Contracts;
-
-// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-
-namespace MotorcycleCompany.Controllers
-{
-    [Route("api/[controller]")]
-    [ApiController]
-    public class CityController : ControllerBase
-    {
-        private readonly IServiceManager _service;
-        public CityController(IServiceManager service)
-        {
-            _service = service;
-        }
-
-        // GET: api/<CityController>
-        [HttpGet]
-        public IActionResult Get()
-        {
-            try
-            {
-                return Ok(_service.CityService.GetAllCities(false));
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, ex.Message);
-            }
-        }
-
-        // GET api/<CityController>/5
-        [HttpGet("{id}")]
-        public string Get(int id)
-        {
-            return "value";
-        }
-
-        // POST api/<CityController>
-        [HttpPost]
-        public IActionResult Post([FromBody] City city)
-        {
-            _service.CityService.CreateCity(city);
-            _service.Save();
-
-            return Ok();
-        }
-
-        // PUT api/<CityController>/5
-        [HttpPut("{id}")]
-        public void Put(int id, [FromBody] string value)
-        {
-        }
-
-        // DELETE api/<CityController>/5
-        [HttpDelete("{id}")]
-        public void Delete(int id)
-        {
-        }
-    }
-}
