+++ MotorcycleCompany/Presentation/Controllers/AgencyController.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using Service.Contracts;
+
+namespace Presentation.Controllers
+{
+    [Route("api/[controller]")]
+    [ApiController]
+    public class AgencyController : ControllerBase
+    {
+        private readonly IServiceManager _service;
+        public AgencyController(IServiceManager service)
+        {
+            _service = service;
+        }
+
+        [HttpGet]
+        public IActionResult Get()
+        {
+            return Ok(_service.AgencyService.GetAllAgenciesDTO(false));
+        }
+    }
+}
